chores: update dependencies (#214)
* Upgrade dependencies and fix compilation errors

* Fix java-time package deprecation

* Update slf4j, sentry, exposed and kotlinx-cli to their latest versions

* Update autohelp to 1.4.0 which supports kord 0.8.0-M7

+++ src/main/kotlin/com/github/devcordde/devcordbot/database/_SnowflakeFields.kt
@@ -59,7 +59,7 @@ object SnowflakeColumnType : ColumnType() {
     override fun valueToDB(value: Any?): Any {
         require(value is Snowflake) { "Value has to be Snowflake" }
 
-        return value.value
+        return value.value.toLong()
     }
 
     override fun valueToString(value: Any?): String = value.toString()
